src/pyproject.toml
[project]
name = "streamlit_app"
authors = [
    {name = "Adrian Jimenez"}
]
description = "Streamlit application"
readme = "README.md"
requires-python = ">=3.12"
version = "0.0.1"
dependencies = [
    "streamlit==1.34.0",
    "langchain==0.1.20",
    "langchain-openai==0.1.7",
    "sentence-transformers==2.7.0",
]

[tool.setuptools]
package-dir = {"streamlit_app" = "streamlit_app"}

[project.optional-dependencies]
dev = [
    "python-dotenv",
    "ruff",
    "mypy",
    "boto3==1.34.107",
    "opensearch-py==2.5.0",
    "requests-aws4auth==1.2.3",
    "pymongo==4.7.2",
    "azure-identity",
    "azure-keyvault-secrets",
]

aws = [
    "boto3==1.34.107",
    "opensearch-py==2.5.0",
    "requests-aws4auth==1.2.3",
]

azure = [
    "azure-identity",
    "azure-keyvault-secrets",
    "pymongo==4.7.2",
]

[tool.mypy]
python_version = "3.12"
strict = true
exclude = [
    '^setup\.py$',
    "venv",
]

[[tool.mypy.overrides]]
module = [
    "setuptools",
    "awsglue.*",
    "dataclasses_json",
    "pyspark.*",
    "awsgluedq.*",
]
ignore_missing_imports = true

[tool.ruff]
select = [
    "E",    # pycodestyle
    "W",    # pycodestyle
    "F",    # pyflake
    "I",    # isort
    "S",    # flake8-bandit
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "D",    # pydocstyle
    "RUF",  # Ruff specific rules
]

exclude = [
    ".git",
    ".idea",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "venv",
    "build",
    "assets",
    "docs/",
]

ignore = [
    "D100",     # Missing docstring in public module
    "D101",     # Missing docstring in public class
    "D102",     # Missing docstring in public method
    "D103",     # Missing docstring in public function
    "D107",     # Missing docstring in __init__
    "D203",     # no-blank-line-before-class (conflict with D211)
    "D213",     # multi-line-summary-second-line (conflict with D212)
    "S608",     # SQL Injection - Should be enabled at some point?
]

line-length = 120

[tool.ruff.per-file-ignores]
"__init__.py" = ["D104"]
"tests/*" = ["S", "D"]

[tool.ruff.isort]
combine-as-imports = true
force-wrap-aliases = true
